Split MapViewer into shader-patches, camera-math, and components (#1089)
Extract the documented GLSL patch (shader-patches.ts), pure orbit/pan/zoom
math (map-camera-math.ts), the instancing renderer (MapInstancedMarkers.tsx),
and the camera controller (MapCamera.tsx). MapViewer becomes the container
(model load, scene setup, ready-gate, chrome). Repoints the test import and
adds map-camera-math unit tests.
